Continuous Teach and Play recording User specified samplerate Variable samplerate (missed samples are accounted for by dt) Switch to turn off csv header in data log conversion
git-svn-id: http://web.barrett.com/svn/btclient/trunk@8 3eb8b4e9-0710-d544-b5f2-d3973d62577e

+++ src/btsystem/control_loop.h
@@ -40,7 +40,7 @@ typedef struct
   SimpleCtl *sc;
 }control_thd_parms;
 
-void start_control_threads(int priority, long sample_period, void *function,void *args);
+void start_control_threads(int priority, double sample_period, void *function,void *args);
 void stop_control_threads();
 void TimerThread();
 void ControlThread(void *data);
